+++ x/gamm/client/cli/flags.go
@@ -5,21 +5,6 @@ import (
 )
 
 const (
-	// Will be parsed to string.
-	FlagPoolFile = "pool-file"
-
-	// Names of fields in pool json file.
-	PoolFileWeights        = "weights"
-	PoolFileInitialDeposit = "initial-deposit"
-	PoolFileSwapFee        = "swap-fee"
-	PoolFileExitFee        = "exit-fee"
-	PoolFileFutureGovernor = "future-governor"
-
-	PoolFileSmoothWeightChangeParams = "lbp-params"
-	PoolFileStartTime                = "start-time"
-	PoolFileDuration                 = "duration"
-	PoolFileTargetPoolWeights        = "target-pool-weights"
-
 	FlagPoolId = "pool-id"
 	// Will be parsed to sdk.Int.
 	FlagShareAmountOut = "share-amount-out"
@@ -38,28 +23,6 @@ const (
 	// Will be parsed to []string.
 	FlagSwapRouteDenoms = "swap-route-denoms"
 )
-
-type createPoolInputs struct {
-	Weights                  string                         `json:"weights"`
-	InitialDeposit           string                         `json:"initial-deposit"`
-	SwapFee                  string                         `json:"swap-fee"`
-	ExitFee                  string                         `json:"exit-fee"`
-	FutureGovernor           string                         `json:"future-governor"`
-	SmoothWeightChangeParams smoothWeightChangeParamsInputs `json:"lbp-params"`
-}
-
-type smoothWeightChangeParamsInputs struct {
-	StartTime         string `json:"start-time"`
-	Duration          string `json:"duration"`
-	TargetPoolWeights string `json:"target-pool-weights"`
-}
-
-func FlagSetCreatePool() *flag.FlagSet {
-	fs := flag.NewFlagSet("", flag.ContinueOnError)
-
-	fs.String(FlagPoolFile, "", "Pool json file path (if this path is given, other create pool flags should not be used)")
-	return fs
-}
 
 func FlagSetJoinPool() *flag.FlagSet {
 	fs := flag.NewFlagSet("", flag.ContinueOnError)
