+++ x/gamm/client/cli/tx.go
@@ -3,12 +3,10 @@ package cli
 import (
 	"errors"
 	"fmt"
-	"time"
 
 	"github.com/spf13/cobra"
 	flag "github.com/spf13/pflag"
 
-	"github.com/osmosis-labs/osmosis/v12/x/gamm/pool-models/balancer"
 	"github.com/osmosis-labs/osmosis/v12/x/gamm/types"
 
 	"github.com/cosmos/cosmos-sdk/client"
@@ -27,7 +25,6 @@ func NewTxCmd() *cobra.Command {
 	}
 
 	txCmd.AddCommand(
-		NewCreatePoolCmd(),
 		NewJoinPoolCmd(),
 		NewExitPoolCmd(),
 		NewJoinSwapExternAmountIn(),
@@ -37,46 +34,6 @@ func NewTxCmd() *cobra.Command {
 	)
 
 	return txCmd
-}
-
-func NewCreatePoolCmd() *cobra.Command {
-	cmd := &cobra.Command{
-		Use:   "create-pool [flags]",
-		Short: "create a new pool and provide the liquidity to it",
-		Long:  `Must provide path to a pool JSON file (--pool-file) describing the pool to be created`,
-		Example: `Sample pool JSON file contents:
-{
-	"weights": "4uatom,4osmo,2uakt",
-	"initial-deposit": "100uatom,5osmo,20uakt",
-	"swap-fee": "0.01",
-	"exit-fee": "0.01",
-	"future-governor": "168h"
-}
-`,
-		Args: cobra.ExactArgs(0),
-		RunE: func(cmd *cobra.Command, args []string) error {
-			clientCtx, err := client.GetClientTxContext(cmd)
-			if err != nil {
-				return err
-			}
-
-			txf := tx.NewFactoryCLI(clientCtx, cmd.Flags()).WithTxConfig(clientCtx.TxConfig).WithAccountRetriever(clientCtx.AccountRetriever)
-
-			txf, msg, err := NewBuildCreateBalancerPoolMsg(clientCtx, txf, cmd.Flags())
-			if err != nil {
-				return err
-			}
-
-			return tx.GenerateOrBroadcastTxWithFactory(clientCtx, txf, msg)
-		},
-	}
-
-	cmd.Flags().AddFlagSet(FlagSetCreatePool())
-	flags.AddTxFlagsToCmd(cmd)
-
-	_ = cmd.MarkFlagRequired(FlagPoolFile)
-
-	return cmd
 }
 
 func NewJoinPoolCmd() *cobra.Command {
@@ -257,105 +214,6 @@ func NewExitSwapShareAmountIn() *cobra.Command {
 	_ = cmd.MarkFlagRequired(FlagPoolId)
 
 	return cmd
-}
-
-func NewBuildCreateBalancerPoolMsg(clientCtx client.Context, txf tx.Factory, fs *flag.FlagSet) (tx.Factory, sdk.Msg, error) {
-	pool, err := parseCreatePoolFlags(fs)
-	if err != nil {
-		return txf, nil, fmt.Errorf("failed to parse pool: %w", err)
-	}
-
-	deposit, err := sdk.ParseCoinsNormalized(pool.InitialDeposit)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	poolAssetCoins, err := sdk.ParseDecCoins(pool.Weights)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	if len(deposit) != len(poolAssetCoins) {
-		return txf, nil, errors.New("deposit tokens and token weights should have same length")
-	}
-
-	swapFee, err := sdk.NewDecFromStr(pool.SwapFee)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	exitFee, err := sdk.NewDecFromStr(pool.ExitFee)
-	if err != nil {
-		return txf, nil, err
-	}
-
-	var poolAssets []balancer.PoolAsset
-	for i := 0; i < len(poolAssetCoins); i++ {
-		if poolAssetCoins[i].Denom != deposit[i].Denom {
-			return txf, nil, errors.New("deposit tokens and token weights should have same denom order")
-		}
-
-		poolAssets = append(poolAssets, balancer.PoolAsset{
-			Weight: poolAssetCoins[i].Amount.RoundInt(),
-			Token:  deposit[i],
-		})
-	}
-
-	poolParams := &balancer.PoolParams{
-		SwapFee: swapFee,
-		ExitFee: exitFee,
-	}
-
-	msg := &balancer.MsgCreateBalancerPool{
-		Sender:             clientCtx.GetFromAddress().String(),
-		PoolParams:         poolParams,
-		PoolAssets:         poolAssets,
-		FuturePoolGovernor: pool.FutureGovernor,
-	}
-
-	if (pool.SmoothWeightChangeParams != smoothWeightChangeParamsInputs{}) {
-		duration, err := time.ParseDuration(pool.SmoothWeightChangeParams.Duration)
-		if err != nil {
-			return txf, nil, fmt.Errorf("could not parse duration: %w", err)
-		}
-
-		targetPoolAssetCoins, err := sdk.ParseDecCoins(pool.SmoothWeightChangeParams.TargetPoolWeights)
-		if err != nil {
-			return txf, nil, err
-		}
-
-		var targetPoolAssets []balancer.PoolAsset
-		for i := 0; i < len(targetPoolAssetCoins); i++ {
-			if targetPoolAssetCoins[i].Denom != poolAssetCoins[i].Denom {
-				return txf, nil, errors.New("initial pool weights and target pool weights should have same denom order")
-			}
-
-			targetPoolAssets = append(targetPoolAssets, balancer.PoolAsset{
-				Weight: targetPoolAssetCoins[i].Amount.RoundInt(),
-				Token:  deposit[i],
-				// TODO: This doesn't make sense. Should only use denom, not an sdk.Coin
-			})
-		}
-
-		smoothWeightParams := balancer.SmoothWeightChangeParams{
-			Duration:           duration,
-			InitialPoolWeights: poolAssets,
-			TargetPoolWeights:  targetPoolAssets,
-		}
-
-		if pool.SmoothWeightChangeParams.StartTime != "" {
-			startTime, err := time.Parse(time.RFC3339, pool.SmoothWeightChangeParams.StartTime)
-			if err != nil {
-				return txf, nil, fmt.Errorf("could not parse time: %w", err)
-			}
-
-			smoothWeightParams.StartTime = startTime
-		}
-
-		msg.PoolParams.SmoothWeightChangeParams = &smoothWeightParams
-	}
-
-	return txf, msg, nil
 }
 
 func NewBuildJoinPoolMsg(clientCtx client.Context, txf tx.Factory, fs *flag.FlagSet) (tx.Factory, sdk.Msg, error) {
